# openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorDto": {
        "properties": {
          "error": {
            "description": "Short error identifier. Usually the standard reason phrase; domain errors use a code, e.g. step_up_required.",
            "example": "Bad Request",
            "type": "string"
          },
          "message": {
            "description": "Human-readable message. An array when request validation failed, one entry per constraint.",
            "example": "Please re-verify your identity to continue.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ]
          },
          "statusCode": {
            "description": "HTTP status code, repeated in the body.",
            "example": 400,
            "type": "integer"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ],
        "type": "object"
      },
...
      "CheckInDto": {
        "properties": {
          "active": {
            "description": "Whether this check-in is active right now",
            "type": "boolean"
          },
          "availabilityEnd": {
            "format": "date-time",
            "type": "string"
          },
          "availabilityStart": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "preparationMinutes": {
            "description": "Minutes the user needs to get ready before heading out",
            "type": "integer"
          },
          "status": {
            "enum": [
              "available",
              "cancelled",
              "expired"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "availabilityStart",
          "availabilityEnd",
          "preparationMinutes",
          "active",
          "createdAt"
        ],
        "type": "object"
      },
...
      "CreateCheckInDto": {
        "properties": {
          "availabilityEnd": {
            "description": "ISO 8601 end of the availability window",
            "example": "2026-07-01T22:00:00.000Z",
            "type": "string"
          },
          "availabilityStart": {
            "description": "ISO 8601 start of the availability window",
            "example": "2026-07-01T18:00:00.000Z",
            "type": "string"
          },
          "latitude": {
            "description": "Latitude of the person's location at check-in, read from the device (−90..90). Required — a location is mandatory to check in. Used only to find nearby people and to pick a nearby venue; never shown to others, who receive a rounded distance instead. This is not a meeting place the person chooses: the venue for a date is selected by the backend and recommended to both people.",
            "example": 52.5186,
            "type": "number"
          },
          "longitude": {
            "description": "Longitude of the person's location at check-in, read from the device (−180..180). Required, together with latitude.",
            "example": 13.3761,
            "type": "number"
          },
          "preparationMinutes": {
            "description": "Minutes the user needs to get ready before heading out (15–60). Defaults to 30 if omitted. Used with travel time to check a date is reachable.",
            "example": 30,
            "type": "integer"
          }
        },
        "required": [
          "availabilityStart",
          "availabilityEnd",
          "latitude",
          "longitude"
        ],
        "type": "object"
      },
...
      "ProfileDto": {
        "properties": {
          "age": {
            "nullable": true,
            "type": "integer"
          },
          "displayName": {
            "nullable": true,
            "type": "string"
          },
          "gender": {
            "nullable": true,
            "type": "string"
          },
          "hiddenFields": {
            "description": "Registry field_ids the user has hidden from their profile. Presentation only — a hidden field does not affect discovery visibility or matching.",
            "example": [
              "age"
            ],
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "id": {
            "type": "string"
          },
          "isComplete": {
            "type": "boolean"
          },
          "isVisible": {
            "type": "boolean"
          },
          "lookingFor": {
            "nullable": true,
            "type": "string"
          },
          "verificationStatus": {
            "enum": [
              "none",
              "pending",
              "verified",
              "rejected"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "displayName",
          "age",
          "gender",
          "lookingFor",
          "isVisible",
          "hiddenFields",
          "isComplete",
          "verificationStatus"
        ],
        "type": "object"
      },
      "ProfileProgressDto": {
        "properties": {
          "displayName": {
            "description": "Step 1. Null when the name has not been given.",
            "nullable": true,
            "type": "string"
          },
          "email": {
            "description": "Step 2. Null when no address has been submitted.",
            "nullable": true,
            "type": "string"
          },
          "emailVerified": {
            "description": "Step 3. False while an address is present but its code has not been confirmed.",
            "type": "boolean"
          },
          "hasDateOfBirth": {
            "description": "Step 4. The date itself is never returned — only whether it is set.",
            "type": "boolean"
          },
          "photoCount": {
            "description": "Confirmed uploads. \"The real you\" step 1 needs four.",
            "type": "integer"
          },
          "promptCount": {
            "description": "\"The real you\" step 2 needs one.",
            "type": "integer"
          }
        },
        "required": [
          "displayName",
          "email",
          "emailVerified",
          "hasDateOfBirth",
          "photoCount",
          "promptCount"
        ],
        "type": "object"
      },
...
      "SetVerificationDto": {
        "properties": {
          "status": {
            "enum": [
              "verified",
              "rejected"
            ],
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
...
      "UpsertProfileDto": {
        "properties": {
          "dateOfBirth": {
            "description": "YYYY-MM-DD; must be 18+",
            "example": "1998-04-23",
            "type": "string"
          },
          "displayName": {
            "example": "Leo",
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "hiddenFields": {
            "description": "Registry field_ids the user does not want displayed on their profile. Replaces the whole set — send the full list, not a delta. Currently accepted: age. Does NOT affect discovery visibility or matching; use isVisible for that.",
            "example": [
              "age"
            ],
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "isVisible": {
            "description": "Whether the profile is visible in discovery",
            "type": "boolean"
          },
          "lookingFor": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Backend API for ShowUp — real-world dating. Source of truth for users, profiles, check-ins, matching, dates, payments, notifications, analytics, safety and admin.",
    "title": "ShowUp Backend API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/profiles/{userId}/verification": {
      "patch": {
        "operationId": "setProfileVerificationStatus",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetVerificationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Request validation failed."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/me/check-ins": {
      "post": {
        "operationId": "createCheckIn",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckInDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckInDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Request validation failed."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "check-ins"
        ]
      }
    },
    "/me/check-ins/active": {
      "get": {
        "operationId": "getActiveCheckIn",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CheckInDto"
                    }
                  ],
                  "nullable": true
                }
              }
            },
            "description": "The active check-in, or null if the user has none."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "check-ins"
        ]
      }
    },
    "/me/check-ins/{id}/cancel": {
      "post": {
        "operationId": "cancelCheckIn",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckInDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Request validation failed."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "check-ins"
        ]
      }
    },
...
    "/me/profile": {
      "get": {
        "operationId": "getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "profiles"
        ]
      },
      "patch": {
        "operationId": "updateProfile",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertProfileDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Request validation failed."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "profiles"
        ]
      },
      "post": {
        "operationId": "createProfile",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertProfileDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Request validation failed."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "profiles"
        ]
      }
    },
    "/me/profile/progress": {
      "get": {
        "operationId": "getProfileProgress",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileProgressDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "profiles"
        ]
      }
    },
    "/me/profile/verification": {
      "post": {
        "operationId": "requestProfileVerification",
        "parameters": [],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "profiles"
        ]
      }
    },
...
  },
  "servers": [],
...
}